requirements.txt
# Core Deep Learning
torch==1.13.1
torchvision==0.14.1
pillow>=9.5.0
numpy>=1.24.0

# Data Analysis
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.2.0

# Image Processing
opencv-python>=4.7.0

# API & Utilities
openai>=1.0.0
requests>=2.31.0
tqdm>=4.65.0

# Web UI
streamlit==1.20.0

# System Monitoring
psutil>=5.9.0

# LLM & Training
transformers>=4.30.0
peft>=0.4.0
datasets>=2.12.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
scipy>=1.10.0
ollama>=0.1.0
qlora>=0.1.0
tensorboard>=2.12.0 